refactor(economy): phase 3 slice 24 compaction and test consolidation (#3939)
Migrate remaining DealMatcher singleFilledDeal closures to firstFilledDeal
pins, merge treasury scenarios into core, fold validator context expectations,
add validatorCtxTimberIron preset, and consolidate resource extractor runners.

Refs #3939
