+++ packages/colonizethis_economy_test_support/lib/src/deal_matcher_scenarios/deal_matcher_core_scenarios.dart
@@ -142,9 +142,9 @@ List<DealMatcherScenario> dealMatcherEmptyAndBasicScenarios() => [
       tradeCapacityByFactionId: {'b': 100},
       pricesByCommodityId: const <CommodityId, double>{},
     ),
-    expect: DealMatchExpectation(
+    expect: const DealMatchExpectation(
       filledDealsLength: 1,
-      singleFilledDeal: (deal) => expect(deal.pricePerUnit, 0.0),
+      firstFilledDeal: FilledDealExpectation(pricePerUnit: 0.0),
     ),
   ),
   DealMatcherScenario.expect(
@@ -300,11 +300,11 @@ List<DealMatcherScenario> dealMatcherBoycottScenarios() => [
     ),
     expect: DealMatchExpectation(
       filledDealsLength: 1,
-      singleFilledDeal: (deal) {
-        expect(deal.sellerFactionId, 'tribeT');
-        expect(deal.buyerFactionId, 'gpD');
-        expect(deal.quantity, 10);
-      },
+      firstFilledDeal: const FilledDealExpectation(
+        sellerFactionId: 'tribeT',
+        buyerFactionId: 'gpD',
+        quantity: 10,
+      ),
       unfilledBidsByFactionId: {
         'gpB': [matcherBid('timber', 10, priority: 1)],
       },
@@ -322,12 +322,12 @@ List<DealMatcherScenario> dealMatcherBoycottScenarios() => [
       },
       tradeCapacityByFactionId: const {'gpB': 100},
     ),
-    expect: DealMatchExpectation(
+    expect: const DealMatchExpectation(
       filledDealsLength: 1,
-      singleFilledDeal: (deal) {
-        expect(deal.buyerFactionId, 'gpB');
-        expect(deal.quantity, 10);
-      },
+      firstFilledDeal: FilledDealExpectation(
+        buyerFactionId: 'gpB',
+        quantity: 10,
+      ),
     ),
     refs: '#3753',
   ),
@@ -507,12 +507,12 @@ List<DealMatcherScenario> dealMatcherLockRecoveryScenarios() => [
       lockRecoverySellerPriorityIds: const {'gp1', 'gp4'},
       treasuryByFactionId: const {'gp1': 100, 'gp4': -50},
     ),
-    expect: DealMatchExpectation(
+    expect: const DealMatchExpectation(
       filledDealsLength: 1,
-      singleFilledDeal: (deal) {
-        expect(deal.sellerFactionId, 'gp4');
-        expect(deal.quantity, 3);
-      },
+      firstFilledDeal: FilledDealExpectation(
+        sellerFactionId: 'gp4',
+        quantity: 3,
+      ),
     ),
     refs: '#2924',
   ),
@@ -584,10 +584,10 @@ List<DealMatcherScenario> dealMatcherSellPriorityScenarios() => [
     ),
     expect: DealMatchExpectation(
       filledDealsLength: 1,
-      singleFilledDeal: (deal) {
-        expect(deal.buyerFactionId, 'gpHigh');
-        expect(deal.quantity, 5);
-      },
+      firstFilledDeal: const FilledDealExpectation(
+        buyerFactionId: 'gpHigh',
+        quantity: 5,
+      ),
       unfilledBidsByFactionId: {
         'gpLow': [matcherBid('timber', 5, priority: 1)],
       },
@@ -609,8 +609,8 @@ List<DealMatcherScenario> dealMatcherSellPriorityScenarios() => [
         'minorM': {'aBuyer': 30, 'zBuyer': 90},
       },
     ),
-    expect: DealMatchExpectation(
-      singleFilledDeal: (deal) => expect(deal.buyerFactionId, 'zBuyer'),
+    expect: const DealMatchExpectation(
+      firstFilledDeal: FilledDealExpectation(buyerFactionId: 'zBuyer'),
     ),
     refs: '#3753',
   ),
@@ -630,7 +630,7 @@ List<DealMatcherScenario> dealMatcherSellPriorityScenarios() => [
       },
     ),
     expect: DealMatchExpectation(
-      singleFilledDeal: (deal) => expect(deal.buyerFactionId, 'gpLow'),
+      firstFilledDeal: const FilledDealExpectation(buyerFactionId: 'gpLow'),
       unfilledBidsByFactionId: {
         'gpHigh': [matcherBid('timber', 5, priority: 1)],
       },
@@ -652,8 +652,8 @@ List<DealMatcherScenario> dealMatcherSellPriorityScenarios() => [
         'minorM': {'gpA': 55, 'gpB': 55},
       },
     ),
-    expect: DealMatchExpectation(
-      singleFilledDeal: (deal) => expect(deal.buyerFactionId, 'gpA'),
+    expect: const DealMatchExpectation(
+      firstFilledDeal: FilledDealExpectation(buyerFactionId: 'gpA'),
     ),
     refs: '#3753',
   ),
@@ -672,8 +672,8 @@ List<DealMatcherScenario> dealMatcherSellPriorityScenarios() => [
         'minorM': {'gpA': 1, 'gpZ': 99},
       },
     ),
-    expect: DealMatchExpectation(
-      singleFilledDeal: (deal) => expect(deal.buyerFactionId, 'gpA'),
+    expect: const DealMatchExpectation(
+      firstFilledDeal: FilledDealExpectation(buyerFactionId: 'gpA'),
     ),
     refs: '#3753',
   ),
@@ -689,8 +689,8 @@ List<DealMatcherScenario> dealMatcherSellPriorityScenarios() => [
       },
       tradeCapacityByFactionId: const {'gpA': 100, 'gpZ': 100},
     ),
-    expect: DealMatchExpectation(
-      singleFilledDeal: (deal) => expect(deal.buyerFactionId, 'gpA'),
+    expect: const DealMatchExpectation(
+      firstFilledDeal: FilledDealExpectation(buyerFactionId: 'gpA'),
     ),
     refs: '#3753',
   ),
@@ -719,5 +719,255 @@ List<DealMatcherScenario> dealMatcherSellPriorityScenarios() => [
       ],
     ),
     refs: '#3753',
+  ),
+];
+
+/// Treasury clamp scenarios from `world_market_deal_matcher_treasury_test.dart`.
+List<DealMatcherScenario> dealMatcherTreasuryScenarios() => [
+  ...dealMatcherTreasuryClampScenarios(),
+  ...dealMatcherTreasuryEdgeCaseScenarios(),
+];
+
+List<DealMatcherScenario> dealMatcherTreasuryClampScenarios() => [
+  DealMatcherScenario.expect(
+    label: 'truncates a single oversized bid to floor(treasury / price)',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'a': [matcherOffer('timber', 10)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('timber', 10)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 100},
+      treasuryBudgetByBuyerFactionId: const {'gp1': 100},
+    ),
+    expect: DealMatchExpectation(
+      filledDealsLength: 1,
+      firstFilledDeal: const FilledDealExpectation(
+        buyerFactionId: 'gp1',
+        quantity: 3,
+        pricePerUnit: 30.0,
+      ),
+      unfilledBidsByFactionId: {
+        'gp1': [matcherBid('timber', 7)],
+      },
+    ),
+    refs: '#3115',
+  ),
+  DealMatcherScenario.expect(
+    label: 'per-buyer running tally exhausts treasury across bids in order',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'sellerA': [matcherOffer('alpha', 5)],
+        'sellerB': [matcherOffer('beta', 5)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('alpha', 5), matcherBid('beta', 5)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 100},
+      treasuryBudgetByBuyerFactionId: const {'gp1': 100},
+      pricesByCommodityId: const {'alpha': 20.0, 'beta': 20.0},
+    ),
+    expect: DealMatchExpectation(
+      filledDealsLength: 1,
+      firstFilledDeal: const FilledDealExpectation(
+        commodityId: 'alpha',
+        quantity: 5,
+      ),
+      unfilledBidsByFactionId: {
+        'gp1': [matcherBid('beta', 5)],
+      },
+    ),
+    refs: '#3115',
+  ),
+  DealMatcherScenario.expect(
+    label: 'negative-treasury buyer treated as zero budget (full suppression)',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'sellerA': [matcherOffer('timber', 10)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('timber', 10)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 100},
+      treasuryBudgetByBuyerFactionId: const {'gp1': -50},
+    ),
+    expect: DealMatchExpectation(
+      filledDealsEmpty: true,
+      unfilledBidsByFactionId: {
+        'gp1': [matcherBid('timber', 10)],
+      },
+    ),
+    refs: '#3115',
+  ),
+  DealMatcherScenario.expect(
+    label: 'FRR pre-pass respects treasury clamp',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'M1': [
+          matcherOffer('timber', 10, originTileKey: 'oldWorld|M1|0|0'),
+        ],
+      },
+      bidsByFactionId: {
+        'gpA': [matcherBid('timber', 10)],
+      },
+      tradeCapacityByFactionId: const {'gpA': 100},
+      treasuryBudgetByBuyerFactionId: const {'gpA': 60},
+      pricesByCommodityId: const {'timber': 20.0},
+      purchasedTileIndex: frrMatcherTestIndex(),
+    ),
+    expect: DealMatchExpectation(
+      filledDealsLength: 1,
+      firstFilledDeal: const FilledDealExpectation(
+        buyerFactionId: 'gpA',
+        quantity: 3,
+        isFirstRightOfRefusalMatch: true,
+      ),
+      unfilledBidsByFactionId: {
+        'gpA': [matcherBid('timber', 10).copyWith(quantity: 7)],
+      },
+    ),
+    refs: '#3115',
+  ),
+];
+
+List<DealMatcherScenario> dealMatcherTreasuryEdgeCaseScenarios() => [
+  DealMatcherScenario.expect(
+    label:
+        'emits exactly one bidPartialFillTreasuryInsufficient note per '
+        'truncated bid (full bid quantity carried in note)',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'a': [matcherOffer('timber', 10)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('timber', 10)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 100},
+      treasuryBudgetByBuyerFactionId: const {'gp1': 100},
+    ),
+    expect: const DealMatchExpectation(
+      activityNotesByCommodityId: {
+        'timber': [
+          MarketActivityNote(
+            kind: MarketActivityNoteKind.bidPartialFillTreasuryInsufficient,
+            factionId: 'gp1',
+            commodityId: 'timber',
+            quantity: 10,
+          ),
+        ],
+      },
+    ),
+    refs: '#3115',
+  ),
+  DealMatcherScenario.expect(
+    label: 'two identical runs produce byte-identical FilledDeal sequences',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'a': [matcherOffer('timber', 10)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('timber', 10)],
+        'gp2': [matcherBid('timber', 10)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 100, 'gp2': 100},
+      treasuryBudgetByBuyerFactionId: const {'gp1': 100, 'gp2': 200},
+    ),
+    deterministicRerun: true,
+    expect: const DealMatchExpectation(),
+    refs: '#3115',
+  ),
+  DealMatcherScenario.expect(
+    label:
+        'zero-price commodity preserves legacy free-fill (no treasury debit)',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'a': [matcherOffer('iron', 5)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('iron', 5)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 100},
+      treasuryBudgetByBuyerFactionId: const {'gp1': 0},
+      pricesByCommodityId: const <CommodityId, double>{},
+    ),
+    expect: const DealMatchExpectation(
+      filledDealsLength: 1,
+      firstFilledDeal: FilledDealExpectation(
+        pricePerUnit: 0.0,
+        quantity: 5,
+      ),
+      activityNotesEmptyForCommodities: ['iron'],
+    ),
+    refs: '#3115',
+  ),
+  DealMatcherScenario.expect(
+    label: 'missing buyer entry in treasury budget treated as zero',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'a': [matcherOffer('timber', 5)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('timber', 5)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 100},
+      treasuryBudgetByBuyerFactionId: const <String, int>{},
+    ),
+    expect: DealMatchExpectation(
+      filledDealsEmpty: true,
+      unfilledBidsByFactionId: {
+        'gp1': [matcherBid('timber', 5)],
+      },
+    ),
+    refs: '#3115',
+  ),
+  DealMatcherScenario.expect(
+    label: 'unaffordable bid at non-zero price emits a note even with zero '
+        'fill quantity',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'a': [matcherOffer('timber', 1)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('timber', 1)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 100},
+      treasuryBudgetByBuyerFactionId: const {'gp1': 10},
+      pricesByCommodityId: const {'timber': 30.0},
+    ),
+    expect: DealMatchExpectation(
+      filledDealsEmpty: true,
+      activityNotesByCommodityId: {
+        'timber': [
+          MarketActivityNote(
+            kind: MarketActivityNoteKind.bidPartialFillTreasuryInsufficient,
+            factionId: 'gp1',
+            commodityId: 'timber',
+            quantity: 1,
+          ),
+        ],
+      },
+    ),
+    refs: '#3115',
+  ),
+  DealMatcherScenario.expect(
+    label: 'cargo clamps tighter than treasury → matchQty falls back to cargo, '
+        'no truncation note emitted',
+    inputs: matcherInputs(
+      offersByFactionId: {
+        'a': [matcherOffer('timber', 10)],
+      },
+      bidsByFactionId: {
+        'gp1': [matcherBid('timber', 10)],
+      },
+      tradeCapacityByFactionId: const {'gp1': 4},
+      treasuryBudgetByBuyerFactionId: const {'gp1': 10_000},
+      pricesByCommodityId: const {'timber': 30.0},
+    ),
+    expect: const DealMatchExpectation(
+      firstFilledDeal: FilledDealExpectation(quantity: 4),
+      activityNotesEmptyForCommodities: ['timber'],
+    ),
+    refs: '#3115',
   ),
 ];
